--- unit-tests/jvm/src/test/scala/org/scalanative/testsuite/javalib/nio/BaseBufferPlatformTest.scala.scala
@@ -1,3 +1,0 @@
-package org.scalanative.testsuite.javalib.nio
-
-trait BaseBufferPlatformTest { self: BaseBufferTest => }
