+++ unit-tests/shared/src/test/require-jdk16/org/scalanative/testsuite/javalib/nio/CharBufferTest.scala
