+++ unit-tests/shared/src/test/require-jdk16/org/scalanative/testsuite/javalib/nio/FloatBufferTest.scala
